+++ XMonad/Util/Timer.hs
@@ -41,7 +41,7 @@ startTimer s = io $ do
     a <- internAtom d "XMONAD_TIMER" False
     allocaXEvent $ \e -> do
          setEventType e clientMessage
-         setClientMessageEvent e rw a 32 (fromIntegral u) currentTime
+         setClientMessageEvent e rw a 32 (fromIntegral u) 0
          sendEvent d rw False structureNotifyMask e
     sync d False
   return u